chore: Composer: fix lock file
The `composer.json` file (as well as the `CONTRIBUTING` doc) indicate that development should use PHP 7.4+, but the project contains a committed `composer.lock` file with dependencies which are not compatible with PHP 7.4.

This fixes the `composer.lock` file to actually contain a PHP 7.4 compatible set of locked dependencies.

Alternatively, it could be considered to remove the `lock` file altogether or to add a `platform-php` setting to it.