GH Actions: re-work the integration tests
This removes three keys from the integration test matrix in favour of adding steps for each of those options to the job.

* The GitHub workspace is cleaned up between each run.
* The Composer cache is removed between each run.

This means that, in effect, the tests are being run in the same way as previously (clean install), but sequentially in one job instead of in parallel in different builds.

This slims down the number of builds per run from 197 to 29.