Update cache action runner
The `actions/cache` package has released version 4.0 which updates the action runner to use Node 20.

Applying this fix gets rid of the following warning which is currently showing in the "Annotations" of workflows using the `ramsey/composer-install` action:
```
Node.js 16 actions are deprecated. Please update the following actions to use Node.js 20: actions/cache@v3. For more information see: https://github.blog/changelog/2023-09-22-github-actions-transitioning-from-node-16-to-node-20/.
```

Note: most action runners have published these types of update in a new major, so that seems to be the best practice. Not sure if this update warrants a new major for `ramsey/composer-install` ?

Refs:
* https://github.com/actions/cache/releases/tag/v4.0.0
* https://github.blog/changelog/2023-09-22-github-actions-transitioning-from-node-16-to-node-20/